Remove FillArrays dependency to reduce import times
It was added in https://github.com/SciML/RecursiveArrayTools.jl/commit/b3ed973b0167f0e60f4b7c4e801af850ce021b17 but in https://github.com/PumasAI/DataInterpolations.jl/issues/129 it has been identified as one of the main import time contributions. It's only necessary for the Zygote rules, and not even ChainRules but specifically `Zygote.@adjoint`, so they are simply moved to a Zygote extension and use FillArrays from Zygote to achieve the same goal.
