+++ src/RecursiveArrayTools.jl
@@ -13,8 +13,6 @@ import ChainRulesCore
 import ChainRulesCore: NoTangent
 import ZygoteRules, Adapt
 
-using FillArrays
-
 import Tables, IteratorInterfaceExtensions
 
 abstract type AbstractVectorOfArray{T, N, A} <: AbstractArray{T, N} end
@@ -41,6 +39,7 @@ import Requires
 @static if !isdefined(Base, :get_extension)
     function __init__()
         Requires.@require Tracker="9f7883ad-71c0-57eb-9f7f-b5c9e6d3789c" begin include("../ext/RecursiveArrayToolsTrackerExt.jl") end
+        Requires.@require Zygote="e88e6eb3-aa80-5325-afca-941959d7151f" begin include("../ext/RecursiveArrayToolsZygoteExt.jl") end
     end
 end
